Rename "get_only_users" to just "get_users"
There are two terms that are important:
- member
- user

User is anyone with an account and a user of the service. It contains
only information about the account.

Member is a person that has a membership and contains information about
the person and it's memberships.

There is no need for the function to be called, get_only_users.
Important notice though, when you get a list of members there user
account information also gets returned with it.

Signed-off-by: Jakob Sinclair <sinclair.jakob@mailbox.org>